refactor(kocolor): migrate FDA feature module to application-specific directory
This commit relocates the FDA feature module from the top-level `:features` directory to the `:applications:kocolor:features` path. This restructuring aligns the module with its primary usage within the KoColor application and updates all internal references, package names, and dependency paths.

- **Project Configuration**:
    - Updated `settings.gradle.kts` and `.idea/gradle.xml` to remove the global `:features:fda` and register the new `:applications:kocolor:features:fda` module.
- **`applications/kocolor/features/fda`**:
    - Re-initialized the module with the new namespace `com.zoewave.probase.kocolor.features.fda`.
    - Migrated the networking layer, including `FdaApi` for searching food/drug enforcements and adverse events, and its corresponding Retrofit data models.
    - Updated `FdaRepository` to handle localized API queries for product recalls and safety reactions.
    - Configured Hilt dependency injection in `FdaModule` to provide singleton instances of the API service.
- **`applications/kocolor/features/cosmetics`**:
    - Updated `CosmeticsViewModel` to import `FdaRepository` from the new package location.
    - Updated `build.gradle.kts` to point to the relocated project dependency.
- **`applications/kocolor/apps/mobile`**:
    - Redirected the mobile application's Gradle dependency to the application-specific FDA module.
