+++ applications/kocolor/apps/mobile/build.gradle.kts
@@ -71,7 +71,7 @@ dependencies {
     implementation(project(":applications:kocolor:features:inventory"))
     implementation(project(":applications:kocolor:features:routines"))
     implementation(project(":applications:kocolor:features:cosmetics"))
-    implementation(project(":features:fda"))
+    implementation(project(":applications:kocolor:features:fda"))
     implementation(project(":features:xr:ar:naillab"))
     implementation(project(":features:xr:ar:facelab"))
