--- features/fda/build.gradle.kts
@@ -1,20 +1,0 @@
-plugins {
-    id("composetemplate.android.library")
-    id("composetemplate.android.hilt")
-    id("composetemplate.kotlin.serialization")
-}
-
-android {
-    namespace = "com.zoewave.probase.features.fda"
-}
-
-dependencies {
-    implementation(project(":core:model"))
-    implementation(project(":core:network"))
-    implementation(project(":applications:kocolor:model"))
-
-    implementation(libs.squareup.retrofit)
-    implementation(libs.squareup.retrofit.converter.gson)
-    implementation(libs.kotlinx.serialization.json)
-    implementation(libs.hilt.android)
-}
