+++ settings.gradle.kts
@@ -43,7 +43,6 @@ include(":features:health")
 include(":features:health:core")
 include(":features:health:cgm")
 include(":features:camera")
-include(":features:fda")
 include(":features:calendar")
 include(":features:xr:ar:naillab")
 include(":features:xr:ar:facelab")
@@ -152,6 +151,7 @@ include(":applications:kocolor:data")
 include(":applications:kocolor:features:analyzer")
 include(":applications:kocolor:features:suggestions")
 include(":applications:kocolor:features:inventory")
+include(":applications:kocolor:features:fda")
 include(":applications:kocolor:features:routines")
 include(":applications:kocolor:features:cosmetics")
 include(":applications:kocolor:features:stitch")
